afs-backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-chroma
chromadb
openai
networkx
python-dotenv